requirements.txt
matplotlib==3.10.7
numpy==2.3.4
Pillow==12.0.0
PyQt5==5.15.11
PyQt5_sip==12.13.0
